% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn.R
\name{classify_nn}
\alias{classify_nn}
\title{Geotag voterfile addresses with 2010 Census block GEOIDs}
\usage{
classify_nn(x)
}
\arguments{
\item{x}{Voterfile dataframe, output from lookup_precincts_nn()}
}
\value{
x with 'PRECINCT_NAME' filled in for blocks with >=1 classified neighbor
}
\description{
We refer to "geotagging" as assigning a Census geography id to an address. This function takes addresses from a voterfile separated into residential address, city and state. It creates a vector of 2010 Census block GEOIDs that can be appended to the voterfile dataframe.
}
