% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn.R
\name{find_neighbors}
\alias{find_neighbors}
\title{Assign a neighborhood to each block}
\usage{
find_neighbors(x, type = "rook")
}
\arguments{
\item{x}{Voterfile dataframe}

\item{type}{Neighborhood type ('rook' or 'queen')}
}
\value{
x with a neighborhood column
}
\description{
Using rook or queen contiguity, create a vector of neighbors for each block. Queen contiguity includes corners while rook does not. Rook contiguity is most commonly used.
}
\examples{
find_neighbors(noble_join, type = 'rook')

}
\keyword{\dontrun}
\keyword{neighbors}
