% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geotag.R
\name{geotag_voterfile}
\alias{geotag_voterfile}
\title{Geotag voterfile addresses with 2010 Census block GEOIDs}
\usage{
geotag_voterfile(x, address_field, city_field, state_field, runtime = TRUE)
}
\arguments{
\item{state}{Postal abbreviation or 2 digit FIPS code of US state \code{state}}

\item{inputParameter2}{County name or 3 digit FIPS code \code{inputParameter2}}
}
\value{
output 'sf' dataframe of 2010 census blocks for the given county
}
\description{
We refer to "geotagging" as assigning a Census geography id to an address. This function takes addresses from a voterfile separated into residential address, city and state. It creates a vector of 2010 Census block GEOIDs that can be appended to the voterfile dataframe.
}
\examples{
load_blocks(state = "OH", county = "Vinton")
load_blocks(state = 39, county = 165) # these will produce the same result
}
\keyword{blocks}
\keyword{census,}
