% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn.R
\name{join_voters_to_blocks}
\alias{join_voters_to_blocks}
\title{Join the geotagged voterfile to 2010 census blocks}
\usage{
join_voters_to_blocks(voters, blocks, block_geoid_voters = "BLOCK_GEOID",
  precinct_name = "PRECINCT_NAME")
}
\arguments{
\item{blocks}{'sf' dataframe of 2010 census blocks. Use load_blocks() to load the blocks for your county}

\item{block_geoid_voters}{Field in x that refers to the census block geoid. Should be called 'BLOCK_GEOID' from geotag_voterfile() output.}

\item{precinct_name}{Field in x that refers to the precinct name. Default= 'PRECINCT_NAME'.}

\item{x}{Geotagged voterfile dataframe, should be output from geotag_voterfile()}
}
\value{
dataframe of blocks with assigned precincts using the most common precinct of the voters in that block
}
\description{
Use this function to assign 2010 census blocks to a particular precinct based on the most common precinct assignment of the voters geotagged there.
}
\examples{
join_voters_to_blocks(noble_voters, noble_blocks, block_geoid_voters = "GEOID10")

}
\keyword{\dontrun}
\keyword{blocks,}
\keyword{voterfile}
