% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_blocks}
\alias{load_blocks}
\title{Load 2010 census blocks for a county}
\usage{
load_blocks(state = "OH", county)
}
\arguments{
\item{state}{Postal abbreviation or 2 digit FIPS code of US state \code{state}}

\item{county}{County name or 3 digit FIPS code \code{county}}
}
\value{
output 'sf' dataframe of 2010 census blocks for the given county
}
\description{
Takes state and county as inputs and returns an 'sf' dataframe of 2010 census blocks using the {tigris} package.
}
\examples{
load_blocks(state = "OH", county = "Vinton")
load_blocks(state = 39, county = 165) # these will produce the same result
}
\keyword{blocks}
\keyword{census,}
