% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_voterfile}
\alias{load_voterfile}
\title{Load the public voterfile for a county}
\usage{
load_voterfile(state = "OH", county)
}
\arguments{
\item{state}{Postal abbreviation of a US state \code{state}}

\item{county}{County name \code{county}}
}
\value{
csv of most current voterfile for the county
}
\description{
Loads a csv of the public voterfile in a county. Currently only supports Ohio counties
}
\examples{
load_voterfile(state = "oh", county = "williams")
}
\keyword{\dontrun}
\keyword{voterfiles}
