% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn.R
\name{lookup_precincts_nn}
\alias{lookup_precincts_nn}
\title{Lookup the precinct names of each block in a neighborhood.}
\usage{
lookup_precincts_nn(x)
}
\arguments{
\item{x}{Voterfile dataframe, output from find_neighbors()}
}
\value{
x with a column of precinct names for each element in the neighborhood
}
\description{
This function translates the vector of neighbors for each census block into a vector of precinct names. This is useful for the nearest neighbor classification.
}
