% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn.R
\name{classify_nn}
\alias{classify_nn}
\title{Geotag voterfile addresses with 2010 Census block GEOIDs}
\usage{
classify_nn(x)
}
\arguments{
\item{x}{Voterfile dataframe}

\item{blocks}{'sf' dataframe of 2010 census blocks. Use load_blocks() to load the blocks for your county}

\item{neighbor}{Type of neighbor relationship to use (rook or queen)}

\item{state_field}{Field in x that refers to the state name}

\item{runtime}{If TRUE, returns the runtime of the geocoding as a message, meaured as start sys.Time() - end sys.Time(). Default = TRUE}
}
\value{
vector of block GEOIDs that can be appended to the voterfile dataframe with dplyr::bind_rows()
}
\description{
We refer to "geotagging" as assigning a Census geography id to an address. This function takes addresses from a voterfile separated into residential address, city and state. It creates a vector of 2010 Census block GEOIDs that can be appended to the voterfile dataframe.
}
\examples{
geotag_voterfile(voterfile, 'RESIDENTIAL_ADDRESS1', 'RESIDENTIAL_CITY', 'RESIDENTIAL_STATE')

}
\keyword{blocks,}
\keyword{geotag}
