% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geotag.R
\name{geotag_voterfile}
\alias{geotag_voterfile}
\title{Geotag voterfile addresses with 2010 Census block GEOIDs}
\usage{
geotag_voterfile(x, address_field, city_field, state_field, runtime = TRUE)
}
\arguments{
\item{x}{Voterfile dataframe \code{x}}

\item{address_field}{Field in x that refers to the street address (house number & street name) \code{address_field}}

\item{city_field}{Field in x that refers to the city name \code{city_field}}

\item{state_field}{Field in x that refers to the state name \code{state_field}}

\item{runtime}{If TRUE, returns the runtime of the geocoding as a message, meaured as start sys.Time() - end sys.Time(). Default = TRUE \code{runtime}}
}
\value{
vector of block GEOIDs that can be appended to the voterfile dataframe with dplyr::bind_rows()
}
\description{
We refer to "geotagging" as assigning a Census geography id to an address. This function takes addresses from a voterfile separated into residential address, city and state. It creates a vector of 2010 Census block GEOIDs that can be appended to the voterfile dataframe.
}
\examples{
geotag_voterfile(voterfile, 'RESIDENTIAL_ADDRESS1', 'RESIDENTIAL_CITY', 'RESIDENTIAL_STATE')
}
\keyword{blocks,}
\keyword{geotag}
